Parse the the cmdline arguments more perfectly
We use getopt to build a standard mechanism to parse command-line
options, so it would be more easily to add new option in the future.
This patch also makes the initrd input to be optional, since it is
not indispensable for kvm-host. README.md is also changed accordingly
for the new command-line interface