Handle serial input using timerfd mechanism
The problem of serial implementation is: If no condition which should be handled happens,
VM will always run inside the loop of KVM_RUN ioctl. To solve the issue, the previous
solution of kvm-host is using another thread to handle serial input. On the other hand, another
solution is using a timer to interrupt main thread periodically, which cause to return from
KVM_RUN ioctl. The latter solution may not certainly be more efficient. For example, although
no POSIX thread is needed for the latter solution, the timer interrupt may cause extra cost if
we return from KVM_RUN ioctl without anything that needs to be handled.

But if we only consider the simplicity, the latter solution seems to be indeed better.
So we change the implementation for such reason now.