Reimplement serial device using epoll and eventfd
After this change, the I/O handler called during VM exit will only need
to deal with the FIFO buffers on the device. The work to interact with
the terminal is offloaded to the worker thread.

The main thread use eventfd to signal the worker thread. The
worker thread uses epoll to monitor stdin, stdout and the eventfd and
will be notified when the state of one of these file descriptors change
and perform the I/O when necessary.

Besides changing the original implementation to use eventfd and epoll,
I also make the emulated device hehave more like a physical device.
The notable change is that the loopback mode is implemented.