Refine FIFO operations of serial device
Instead of sending signal to notify the main thread to read more bytes
into rx_buf, just do it in the worker thread.

When rx_buf is full but stdin is still readable, the worker thread
waits on the condition variable. The main thread will wake the worker
thread when the level of the rx_buf decreases to its half size.

Because rx_buf, iir, ier and lsr may be accessed by multiple thread, use
mutex and atomic operations to avoid the race condition.

Close #15