Remove Initial Value for vq->info.notify_off in virtio-net
This commit removes the initialization of `vq->info.notify_off`
value in the virtio-net implementation. The reason for this
change is that during the execution of RX and TX enable
callback functions, the `vm_ioeventfd_register` function
registers the ioeventfd at the same address repeatly.
Therefore, the notify_off value is now initialized directly
within the virtqueue enable callback functions.