Implement vm_enable_net API
The vm_enable_net function to facilitate the integration
of additional I/O devices in the future. The function initializes
the virtio-net device by setting up the tap device using virtio_net_init
and subsequently registers the virtio-net device with the PCI bus.
This function is invoked during the VM initialization session.